requirements.txt
mcp
PyPDF
langchain-cerebras
python-dotenv 
